Move Queue to root module
